Keep the file-preview classifier out of electron's import path
The publish-npm job installs root deps only, then runs the prepublishOnly gate
(typecheck + test + smoke-pack). tests/file-preview-classification.test.ts
imported isTextLikeForPreview from files-handler.ts, which imports electron at
module top-level -- and with no app/node_modules, vite can't even resolve the
bare 'electron' specifier at transform time, so the suite fails to load (the
existing vi.mock can't help; it applies after resolution). That tanked the
0.5.0 npm publish. Split the pure classifier into its own electron-free module
and point the test at it. Validated with a root-only npm ci: the full
prepublishOnly chain is green.
