+++ app/src/main/files-handler.ts
@@ -9,6 +9,7 @@
 
 import { ipcMain, BrowserWindow } from "electron";
 import { createIdempotentIpc } from "./ipc-registry.js";
+import { isTextLikeForPreview } from "./file-preview-classification.js";
 import * as fs from "node:fs";
 import * as fsp from "node:fs/promises";
 import * as path from "node:path";
@@ -43,80 +44,6 @@ const MAX_READ_BYTES = 5 * 1024 * 1024; // 5 MB — full read up to here
 const MAX_PREVIEW_BYTES = 1024 * 1024 * 1024; // 1 GB — hard refuse above this
 const PREVIEW_LINE_COUNT = 10;
 const PREVIEW_BYTE_BUDGET = 64 * 1024; // 64 KB cap on the head we read
-
-// Mirrors the renderer's TEXT_EXTS in file-viewer.ts -- the head-preview
-// path only makes sense for files the renderer would draw as text. Returning
-// 64 KB of head bytes for an image / pdf / binary would just produce a
-// broken <img> or corrupted pdf in the renderer, with no useful signal to
-// the user. Files outside this set in the (5 MB, 1 GB] band get the same
-// "too large" rejection they got before head-preview existed.
-const TEXT_PREVIEW_EXTS = new Set([
-  ".md",
-  ".txt",
-  ".log",
-  ".rst",
-  ".py",
-  ".js",
-  ".ts",
-  ".tsx",
-  ".jsx",
-  ".sh",
-  ".rb",
-  ".pl",
-  ".r",
-  ".go",
-  ".rs",
-  ".json",
-  ".jsonl",
-  ".yml",
-  ".yaml",
-  ".toml",
-  ".ini",
-  ".cfg",
-  ".conf",
-  ".xml",
-  ".html",
-  ".htm",
-  ".css",
-  ".csv",
-  ".tsv",
-  ".tab",
-  ".tabular",
-  ".fa",
-  ".fasta",
-  ".fna",
-  ".faa",
-  ".ffn",
-  ".fastq",
-  ".fq",
-  ".fastqsanger",
-  ".fastqillumina",
-  ".fastqsolexa",
-  ".fastqcssanger",
-  ".vcf",
-  ".bed",
-  ".interval",
-  ".bedgraph",
-  ".wig",
-  ".gff",
-  ".gff3",
-  ".gtf",
-  ".sam",
-  ".pdb",
-  ".cif",
-  ".nwk",
-  ".newick",
-  ".tree",
-  ".phy",
-  ".phylip",
-]);
-
-export function isTextLikeForPreview(name: string): boolean {
-  const dot = name.lastIndexOf(".");
-  // No extension -- the renderer treats these as text (READMEs, configs).
-  if (dot <= 0) return true;
-  return TEXT_PREVIEW_EXTS.has(name.slice(dot).toLowerCase());
-}
 
 /**
  * Clamp a user-supplied path to the current cwd. Throws if it escapes.
